add melon v2 to melon adapter
